users routes order in progress

+++ routes/users.js
@@ -6,6 +6,10 @@ import c from '../config.js'
 
 export const users = Router()
 
+const range = {
+  cost: 9999,
+  role_id: 10
+}
 const FF = Db.model('follower_following')
 const UM = Db.model('user_major')
 
@@ -21,7 +25,6 @@ const UM = Db.model('user_major')
  * @apiParam (Query String) {string} [province] Filtered by the province a user is living in, check "Provinces list"
  * @apiParam (Query String) {string} [countries] Filtered by the country a user is living in, check "Countries list"
  * @apiParam (Query String) {boolean=0,1} [active=0,1] Filtered by if a user wished to be found
- * @apiParam (Query String) {boolean=0,1} [available=0,1] Filtered by if a user is opened for booking
  * @apiParam (Query String) {string=DESC, ASC} [cost] Sorting by cost
  * @apiParam (Query String) {integer} [page=1] Pagination
  * @apiParamExample {json} Request-example:
@@ -30,22 +33,9 @@ const UM = Db.model('user_major')
  */
 users.get('/', async (ctx) => {
   try {
-    const filters = [
-      'id', 'mobilephone', 'role_id', 'gender',
-      'place', 'province', 'city', 'country',
-      'active', 'available'
-    ]
+    const filters = ['id', 'mobilephone', 'role_id', 'gender', 'place', 'province', 'city', 'country', 'active']
     const qs = General.parseQuerystring(ctx.request.querystring)
     let filter = General.objToObjGroupsInArr(qs, filters)
-    let sorting = []
-
-    for (let key in qs) {
-      // ASC as default order
-      if (['cost'].indexOf(key) !== -1) {
-        qs[key] = qs[key] === 'DESC' ? 'DESC' : 'ASC'
-        sorting.splice(0, 0, [key, qs[key]])
-      }
-    }
 
     // this part is for majors filtering
     // if majors is given in the querystring then do the follow
@@ -60,7 +50,6 @@ users.get('/', async (ctx) => {
       limit: c.queryLimit,
       offset: General.getOffset(qs.page, c.queryLimit),
       where: { $and: filter },
-      order: sorting,
       attributes: { exclude: ['password'] }
     })
 
@@ -69,15 +58,16 @@ users.get('/', async (ctx) => {
       await UserUtils.addFields(data[i].dataValues, data[i].dataValues.id)
     }
 
-    // Reorder
+    // Order for teacher 
     if (qs.role_id === '2') {
       data.map(each => {
         each.dataValues.weight = UserUtils.generalOrder(each.dataValues)
       })
+      data.sort((a, b) => b.dataValues.weight - a.dataValues.weight)
     }
-    data.sort((a, b) => b.dataValues.weight - a.dataValues.weight)
 
-    General.simpleSend(ctx, data)
+    ctx.status = 200
+    ctx.body = General.prettyJSON(data)
   } catch (err) {
     General.logError(ctx, err)
   }
@@ -95,8 +85,9 @@ users.get('/:id', async (ctx) => {
 
     if (data) {
       await UserUtils.addFields(data.dataValues, id)
-      
-      General.simpleSend(ctx, data)
+      ctx.status = 200
+      ctx.body = General.prettyJSON(data)
+
     } else {
       ctx.status = 404
     }
@@ -134,7 +125,8 @@ users.get('/:id/students', async (ctx) => {
       await UserUtils.addFields(data[i].dataValues, data[i].dataValues.id)
     }
 
-    General.simpleSend(ctx, data)
+    ctx.status = 200
+    ctx.body = General.prettyJSON(data)
   } catch (err) {
     General.logError(ctx, err)
   }
@@ -169,7 +161,8 @@ users.get('/:id/teachers', async (ctx) => {
       await UserUtils.addFields(data[i].dataValues, data[i].dataValues.id)
     }
 
-    General.simpleSend(ctx, data)
+    ctx.status = 200
+    ctx.body = General.prettyJSON(data)
   } catch (err) {
     General.logError(ctx, err)
   }
@@ -200,7 +193,8 @@ users.get('/:id/followers', async (ctx) => {
       await UserUtils.addFields(data[i].dataValues, data[i].dataValues.id)
     }
 
-    General.simpleSend(ctx, data)
+    ctx.status = 200
+    ctx.body = General.prettyJSON(data)
   } catch (err) {
     General.logError(ctx, err)
   }
@@ -231,7 +225,8 @@ users.get('/:id/followings', async (ctx) => {
       await UserUtils.addFields(data[i].dataValues, data[i].dataValues.id)
     }
 
-    General.simpleSend(ctx, data)
+    ctx.status = 200
+    ctx.body = General.prettyJSON(data)
   } catch (err) {
     General.logError(ctx, err)
   }
@@ -243,12 +238,9 @@ users.get('/:id/followings', async (ctx) => {
  * @apiParam {string} name User name
  * @apiParam {string} mobilephone User mobilephone number
  * @apiParam {string} password User passowrd
+ * @apiParam {string=1,2,3} role_id User's role
+ * @apiParam {boolean} active If user can be seached or not
  * @apiParamExample {json} Request-example:
- *  {
- *    "name": "Tony",
- *    "mobilephone": "12345678901",
- *    "password": "000000" 
- *  }
  * @apiSuccess (201) {object} void The newly created user object
  */
 users.put('/', async (ctx) => {
@@ -284,28 +276,37 @@ users.post('/:id', async (ctx) => {
   try {
     const user_id = ctx.params.id
     let values = ctx.request.body
+    const isOutRange = General.checkRange(range, values)
 
-    await UM.destroy({
-      where: { user_id }
-    })
-    values.majors.map(async major_id => {
-      await UM.create({ user_id, major_id })
-    })
-    await Db.sync()
+    if (isOutRange) {
+      ctx.status = 416
+      ctx.body = isOutRange
 
-    let data = await UserUtils.getOneUser(user_id)
-    // Delete majors because it's not updated here
-    delete values.majors
-    data = await data.update(values)
+    } else {
+      if (values.majors) {
+        await UM.destroy({ where: { user_id } })
+        values.majors.map(async major_id => {
+          await UM.create({ user_id, major_id })
+        })
+        await Db.sync()
+      }
 
-    // do not send password to client
-    delete data.dataValues.password
+      let data = await UserUtils.getOneUser(user_id)
+      // Delete majors because it's not updated here
+      delete values.majors
+      data = await data.update(values)
 
-    // Add majors list
-    const majors = await Db.model('user_major').findAll({ where: { user_id } })
-    data.dataValues.majors = majors.map(each => each.major_id)
+      // do not send password to client
+      delete data.dataValues.password
 
-    General.simpleSend(ctx, data)
+      // Add majors list
+      const majors = await Db.model('user_major').findAll({ where: { user_id } })
+      data.dataValues.majors = majors.map(each => each.major_id)
+
+      ctx.status = 200
+      ctx.body = General.prettyJSON(data)
+    }
+
   } catch (err) {
     General.logError(ctx, err)
   }
